chore(v3): fold core_components.ex into HuddlzWeb.Components.*
Split the four survivors of core_components.ex into their own modules
under lib/huddlz_web/components/ to match the per-file pattern already
used by Button, Card, Chip, etc. Wires them into HuddlzWeb.Components
so `use HuddlzWeb.Components` brings every function component into
scope at once and the top-level `import HuddlzWeb.CoreComponents`
in huddlz_web.ex can be dropped.

- HuddlzWeb.Components.Flash    — flash/1, show/1, hide/1
- HuddlzWeb.Components.Icon     — icon/1
- HuddlzWeb.Components.Avatar   — avatar/1 + private helpers
- HuddlzWeb.Components.Modal    — modal/1, show_modal/2, hide_modal/2
